consumer-requirements.txt
pika==1.1.0
mysql-connector==2.2.9